# archive/tar: improve parseOctal performance